server/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.48
pymysql==1.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
aiofiles==25.1.0
websockets==14.1
bcrypt==4.0.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyJWT==2.10.1
